Roughed out an fkfs_log implementation.
This basically buffers a blocks worth of log that can be printed to and
then writes that block to an fkfs. In many situations this will be a bulk
of the data written and it's nice to avoid fragmentation by buffering in
this way. It does mean that very early logs could just get blown away
and so there's probably some value in starting new log files at startup.