Fix some iterating file bugs.
Big one here is the missing of data that doesn't fall at the beginning
because of the math around moving the data pointer. Also adds a ton of
logging code. Another addition is if we've got 10 blocks with no data we
bail, putting an upper limit on our CPU consumption.