Fix iteration bug. FK-209
When we began sending both the data and log files to the server things started to go to shit. Suddenly the iteration of the data file would end prematurely, hitting the last block but not finding all of the data. It took some digging but I eventually figured this out.

What was happening was the iteration of the log would work fine and then we would truncate the log file. This would cause the version on the log file to change. Which means when we would iterate over the data file, we'd come across those old blocks and they would fail the CRC check. The old behavior was to skip to the next block.

The fix was to keep scanning within a block if just the CRC failed, and things like the sizes were accurate.